moving UI tests to ui
