--- webtau-feature-testing/examples/Basic.groovy
@@ -1,3 +1,0 @@
-import static com.twosigma.webtau.WebTauDsl.open
-
-open("/relative-url")
