--- webtau-feature-testing/examples/rest/headers/restGet.groovy
@@ -1,10 +1,0 @@
-package rest.headers
-
-import static com.twosigma.webtau.WebTauDsl.http
-import static com.twosigma.webtau.WebTauGroovyDsl.scenario
-
-scenario("simple get") {
-    http.get("/weather") {
-        temperature.should == 88
-    }
-}
