--- webtau-feature-testing/examples/rest/headers/webtau.cfg
@@ -1,5 +1,0 @@
-import rest.headers.auth.Auth
-
-url = "http://localhost:8080"
-
-httpHeaderProvider = Auth.&authHeader
