--- webtau-feature-testing/examples/rest/report/webtau.cfg
@@ -1,5 +1,0 @@
-import rest.report.Report
-
-url = "http://localhost:8080"
-
-reportGenerator = Report.&generateReport
