+++ webtau-feature-testing/examples/scenarios/rest/docArtifacts.cfg
