+++ webtau-feature-testing/examples/scenarios/rest/report/Report.groovy
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2018 TWO SIGMA OPEN SOURCE, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package scenarios.rest.report
+
+import com.twosigma.webtau.console.ConsoleOutputs
+import com.twosigma.webtau.console.ansi.Color
+import com.twosigma.webtau.report.ReportTestEntries
+
+import static com.twosigma.webtau.WebTauDsl.cfg
+
+class Report {
+    static void generateReport(ReportTestEntries entries) {
+        def reportPath = cfg.workingDir.resolve('report.txt')
+
+        ConsoleOutputs.out('generating report: ', Color.PURPLE, reportPath)
+        reportPath.toFile().text = entries.size()
+    }
+}
