+++ webtau-feature-testing/examples/scenarios/rest/springboot/listContain.groovy
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2018 TWO SIGMA OPEN SOURCE, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package scenarios.rest.springboot
+
+import static com.twosigma.webtau.Ddjt.contain
+import static com.twosigma.webtau.WebTauDsl.http
+import static com.twosigma.webtau.WebTauGroovyDsl.scenario
+
+scenario("list Customers and assert that it contains a specified entry") {
+    http.get("/customers/search/findAllByOrderByLastName") {
+        _embedded.customers.should contain([firstName: 'FN1', lastName: 'LN1'])
+    }
+}
