+++ webtau-feature-testing/examples/scenarios/ui/byNumber.groovy
@@ -1,11 +1,10 @@
-package ui
+package scenarios.ui
 
 import static com.twosigma.webtau.WebTauDsl.$
-import static com.twosigma.webtau.WebTauDsl.getCfg
 import static com.twosigma.webtau.WebTauDsl.open
 import static com.twosigma.webtau.WebTauGroovyDsl.scenario
 
-scenario("accessing custom config value") {
+scenario("filter by number") {
     open("/finders-and-filters")
-    $("#menu ul li a").get(2).should == cfg.userName
+    $("#menu ul li a").get(2).should == 'orders'
 }
