+++ webtau-feature-testing/examples/scenarios/ui/byRegexp.groovy
@@ -1,4 +1,4 @@
-package ui
+package scenarios.ui
 
 import static com.twosigma.webtau.WebTauDsl.$
 import static com.twosigma.webtau.WebTauDsl.open
