+++ webtau-feature-testing/examples/scenarios/ui/cfgAccess.groovy
@@ -1,10 +1,11 @@
-package ui
+package scenarios.ui
 
 import static com.twosigma.webtau.WebTauDsl.$
+import static com.twosigma.webtau.WebTauDsl.getCfg
 import static com.twosigma.webtau.WebTauDsl.open
 import static com.twosigma.webtau.WebTauGroovyDsl.scenario
 
-scenario("filter by number") {
+scenario("accessing custom config value") {
     open("/finders-and-filters")
-    $("#menu ul li a").get(2).should == 'orders'
+    $("#menu ul li a").get(2).should == cfg.userName
 }
