+++ webtau-feature-testing/examples/scenarios/ui/should.groovy
@@ -1,4 +1,4 @@
-package ui
+package scenarios.ui
 
 import static com.twosigma.webtau.WebTauGroovyDsl.scenario
 import static pages.Pages.getSearch
