+++ webtau-feature-testing/examples/scenarios/ui/waitToNot.groovy
@@ -1,13 +1,11 @@
-package ui
+package scenarios.ui
 
 import static com.twosigma.webtau.WebTauGroovyDsl.scenario
 import static pages.Pages.getSearch
 
 scenario("""Synchronization point with dynamic UI.
-Executes matcher multiple time until it matches or time runs out.""") {
+Executes negated matcher multiple time until it matches or time runs out.""") {
     search.open()
-
     search.submit(query: "search this")
-    search.numberOfResults.waitTo == 2
+    search.numberOfResults.waitToNot == 1
 }
-
