+++ webtau-feature-testing/examples/scenarios/ui/windowSize.cfg
