--- webtau-feature-testing/examples/simpleRest.groovy
@@ -1,8 +1,0 @@
-import static com.twosigma.webtau.WebTauDsl.http
-import static com.twosigma.webtau.WebTauGroovyDsl.scenario
-
-scenario("simple get") {
-    http.get("/weather") {
-        temperature.should == 88
-    }
-}
